# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a calculator server implementing a Hexagonal Architecture.",
        "title": "Calculator HTTP API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/add": {
            "post": {
                "description": "Adds operand A to operand B.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binary Operations"
                ],
                "summary": "Addition",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BinaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Error like Invalid JSON, NaN/Inf, or Overflow",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/divide": {
            "post": {
                "description": "Divides operand A by operand B. Returns a 400 Error in case of division by zero.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binary Operations"
                ],
                "summary": "Division",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BinaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exponentiate": {
            "post": {
                "description": "Raises base to the power of exponent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binary Operations"
                ],
                "summary": "Exponentiation",
                "parameters": [
                    {
                        "description": "Base and exponent",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BinaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Error like Invalid JSON, NaN/Inf, or Overflow",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/multiply": {
            "post": {
                "description": "Multiplies operand A by operand B.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binary Operations"
                ],
                "summary": "Multiplication",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BinaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/negate": {
            "post": {
                "description": "Returns the negative value of the input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unary Operations"
                ],
                "summary": "Negation (Sign Change)",
                "parameters": [
                    {
                        "description": "Single operand",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/percentage": {
            "post": {
                "description": "Returns the percentage of the given value (value / 100).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unary Operations"
                ],
                "summary": "Percentage",
                "parameters": [
                    {
                        "description": "Single operand",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/squareroot": {
            "post": {
                "description": "Returns the square root of the given value. Returns a 400 Error for negative values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unary Operations"
                ],
                "summary": "Square Root",
                "parameters": [
                    {
                        "description": "Single operand",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subtract": {
            "post": {
                "description": "Subtracts operand B from operand A.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Binary Operations"
                ],
                "summary": "Subtraction",
                "parameters": [
                    {
                        "description": "Operands",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BinaryOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calculation result",
                        "schema": {
                            "$ref": "#/definitions/models.CalculationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BinaryOperationRequest": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                }
            }
        },
        "models.CalculationResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "number"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.UnaryOperationRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "number"
                }
            }
        }
    }
}